requirements.txt
altgraph==0.17.2
certifi==2021.10.8
future==0.18.2
mysql-connector-python==8.0.27
pefile==2021.9.3
protobuf==3.19.1
pyinstaller==4.7
pyinstaller-hooks-contrib==2021.5
python-dateutil==2.8.2
pywin32-ctypes==0.2.0
six==1.16.0
urllib3==1.26.8
